scrub snapping based on ctrl held down, organize timeline scrubto[Count/MS]

+++ src/main/java/org/emrick/project/TimelineListener.java
@@ -8,5 +8,5 @@ public interface TimelineListener {
      * Called when the timeline is scrubbed to a specific count
      * @param count The count position that was scrubbed to
      */
-    void onTimelineScrub(int count);
+    void onTimelineScrub(double count);
 }
